Move server interpreter & interceptor code to a separate module
