+++ docs/openapi-docs/src/main/scala/sttp/tapir/docs/openapi/EndpointInputToDecodeFailureOutput.scala
@@ -2,8 +2,9 @@ package sttp.tapir.docs.openapi
 
 import sttp.model.MediaType
 import sttp.tapir.internal._
-import sttp.tapir.server.interceptor.decodefailure.DefaultDecodeFailureHandler
 import sttp.tapir.{Codec, CodecFormat, EndpointIO, EndpointInput, SchemaType}
+
+import scala.annotation.tailrec
 
 private[openapi] object EndpointInputToDecodeFailureOutput {
   def defaultBadRequestDescription(input: EndpointInput[_]): Option[String] = {
@@ -29,8 +30,29 @@ private[openapi] object EndpointInputToDecodeFailureOutput {
       codec.schema.schemaType != SchemaType.SString()
 
   private def badRequestDescription(fallibleBasicInputs: Vector[EndpointInput.Basic[_]]) =
-    fallibleBasicInputs
-      .map(input => DefaultDecodeFailureHandler.FailureMessages.failureSourceMessage(input))
-      .distinct
-      .mkString(", ")
+    fallibleBasicInputs.map(failureSourceMessage).distinct.mkString(", ")
+
+  /** Describes the source of the failure: in which part of the request did the failure occur. Currently the same as
+    * `DefaultDecodeFailureHandler.FailureMessages.failureSourceMessage`
+    */
+  @tailrec
+  def failureSourceMessage(input: EndpointInput[_]): String =
+    input match {
+      case EndpointInput.FixedMethod(_, _, _)      => s"Invalid value for: method"
+      case EndpointInput.FixedPath(_, _, _)        => s"Invalid value for: path segment"
+      case EndpointInput.PathCapture(name, _, _)   => s"Invalid value for: path parameter ${name.getOrElse("?")}"
+      case EndpointInput.PathsCapture(_, _)        => s"Invalid value for: path"
+      case EndpointInput.Query(name, _, _)         => s"Invalid value for: query parameter $name"
+      case EndpointInput.QueryParams(_, _)         => "Invalid value for: query parameters"
+      case EndpointInput.Cookie(name, _, _)        => s"Invalid value for: cookie $name"
+      case _: EndpointInput.ExtractFromRequest[_]  => "Invalid value"
+      case a: EndpointInput.Auth[_, _]             => failureSourceMessage(a.input)
+      case _: EndpointInput.MappedPair[_, _, _, _] => "Invalid value"
+      case _: EndpointIO.Body[_, _]                => s"Invalid value for: body"
+      case _: EndpointIO.StreamBodyWrapper[_, _]   => s"Invalid value for: body"
+      case EndpointIO.Header(name, _, _)           => s"Invalid value for: header $name"
+      case EndpointIO.FixedHeader(name, _, _)      => s"Invalid value for: header $name"
+      case EndpointIO.Headers(_, _)                => s"Invalid value for: headers"
+      case _                                       => "Invalid value"
+    }
 }
