+++ server/core/src/main/scala/sttp/tapir/server/interceptor/DecodeSuccessContext.scala
