+++ server/core/src/main/scala/sttp/tapir/server/interceptor/SecurityFailureContext.scala
