+++ server/core/src/main/scala/sttp/tapir/server/interceptor/content/UnsupportedMediaTypeInterceptor.scala
