+++ server/core/src/main/scala/sttp/tapir/server/interceptor/cors/CORSConfig.scala
