+++ server/core/src/main/scala/sttp/tapir/server/interceptor/exception/ExceptionHandler.scala
