+++ server/core/src/main/scala/sttp/tapir/server/interceptor/reject/RejectHandler.scala
