+++ server/core/src/main/scala/sttp/tapir/server/interceptor/reject/RejectInterceptor.scala
