+++ server/core/src/main/scala/sttp/tapir/server/interpreter/DecodeBasicInputs.scala
