+++ server/core/src/main/scala/sttp/tapir/server/interpreter/InputValue.scala
