+++ server/core/src/test/scala/sttp/tapir/server/interceptor/cors/CORSInterceptorTest.scala
